+++ crates/selector4nix-core/src/infrastructure/config/mod.rs
@@ -5,8 +5,8 @@ mod general_raw;
 
 pub use credential_parsed::{AppCredential, AppCredentialEntry};
 pub use general_parsed::{
-    AppConfiguration, CacheConfiguration, CacheInfoConfiguration,
+    AppConfiguration, BandwidthProbeConfiguration, CacheConfiguration, CacheInfoConfiguration,
     CloudflareCacheProxyConfiguration, CloudflareOptimizationConfiguration,
-    FastlyOptimizationConfiguration, NetworkConfiguration, ProxyConfiguration, ServerConfiguration,
-    SubstituterConfiguration,
+    ExternalIpListConfiguration, FastlyOptimizationConfiguration, NetworkConfiguration,
+    ProxyConfiguration, ServerConfiguration, SubstituterConfiguration,
 };