+++ crates/selector4nix-core/src/infrastructure/provider/mod.rs
@@ -1,6 +1,7 @@
 mod derivation_log_provider;
 mod endpoint_client_pool;
 mod endpoint_probing_provider;
+mod external_ip_list_provider;
 mod nar_directory_provider;
 mod nar_info_provider;
 mod nar_stream_provider;
@@ -9,6 +10,7 @@ mod substituter_probing_provider;
 pub use derivation_log_provider::ReqwestDerivationLogProvider;
 pub use endpoint_client_pool::{EndpointClientPool, EndpointClientSet};
 pub use endpoint_probing_provider::{EndpointProbingProvider, ProbeEndpointError};
+pub use external_ip_list_provider::ExternalIpListProvider;
 pub use nar_directory_provider::ReqwestNarDirectoryProvider;
 pub use nar_info_provider::ReqwestNarInfoProvider;
 pub use nar_stream_provider::ReqwestNarStreamProvider;